Move validate_restrictions and validate_door_types to BuildTagsManager
